added players list dialog
